Fix potential loss of force_draw when grouping MultiProgress updates
MultiProgress was only passing on the force_draw flag to its draw_target if it
is set on the last ProgressDrawState in a group.

Fix this by handling force_draw similarly to orphan_lines: initialize it outside
the loop, accumulate (set it to true if necessary) as we receive updates, clear
it (set it back to false) after drawing.

I think it is currently not possible to actually hit this bug: the only places
producing a ProgressDrawState with force_draw set are println() and
MultiProgress itself, using println() always forces a draw because orphan_lines
is nonzero, and it looks like it is not currently possible to nest a
MultiProgress in another MultiProgress. I only noticed this because I was making
an unrelated change to this code. But we might as well fix it to avoid future
problems.