Fix fine-grained progress bars
Commit 3cbd533da3eda9162da8955ead6218a022b0604b shadowed the floating-point
`fill` variable with the integer number of entirely-filled characters in the
progress bar. We still need the floating-point progress to pick the fine-grained
progress character: after that commit, `(fill * n) % n` is always 0, so we
always use the last fine-grained character.

Fix this by using a different name for the integer progress.

While I'm there, try to make the function easier to read:

- Use `str::repeat`.
- Rename some variables and add comments.
- Remove the unnecessary `pct` variable.
- Skip the selection of a fine-grained character when we have exactly one.
- Use `fill.fract()` instead of modulo to discard the integer part of progress.
  This should make it difficult to reintroduce the same bug.

Fixes #206