Estimator: Fix panic when rewinding position
In some situations, we can get a panic like this:

```
The application panicked (crashed).
Message:  attempt to subtract with overflow
Location: /home/runner/.cargo/registry/src/github.com-1ecc6299db9ec823/indicatif-0.17.0-rc.9/src/state.rs:339
```

This could happen if we set the progress of a meter to a lower value than it was previously. This commit fixes the panic that would happen in such cases.