Use buffer for hold_max field in ProgressIter
On the first seek it gets enabled, displaying the progress as the max
of the last 10 updates.
If there ever are more than 5 consecutive reads and writes without seek
it gets disabled again, keeping the performance impact low.

This also fixes a pre-existing logic error in the async BufRead impl.