Fix duration after finish
When initializing progress bar with style that exposes template variable
`{duration}`, the duration is set to 0 after progress bar finishes.

This is is unexpected and makes progress bar styles other than
'ProgressFinish::AndClear' display wrong duration.

This will add Duration parameter to `crate::state::State::Done*`
variants and uses it to display duration if curret state of the progress
bar is one of the Done states.

Fixes #557